+++ src/providers/openfoodfacts/index.ts
@@ -1,2 +1,4 @@
+export * from "./types.js";
+export * from "./client.js";
 export * from "./constants.js";
 export * from "./product.js";